Do not initially share the SD card over USB
Recent changes to Marlin mean that when using a shared SD card Marlin will mount it at boot time. This fix changes the HAL startup code so that it no longer initially makes the SD card available over USB. Without this a host system will see the card, start reading it and then find it is no longer available (which causes errors). The card can be made availabe via USB to a host system by releasing the card either via the LCD or via gcode M22